# Serialize less memory