This updates the `Serialize` function 

Previously we were collecting all row groups in the entire table block, and using the parquet merge function to merge all the row groups into a single row group. This causes a large memory spike because we would load at least a page for every row group (or maybe 2 if we're async reading the pages) into memory to perform the heap sort of all the row groups.

Instead of doing that; we can leverage the fact that Granules are already sorted in relation to each other and instead merge each Granules row groups into a single one, and write that directly to the final file. The benchmark shows some considerable gains in memory, and speed.

```
thor@thors-MacBook-Pro frostdb % benchstat before.txt after.txt
name           old time/op    new time/op    delta
_Serialize-10    3.38ms ± 3%    1.42ms ± 0%  -58.13%  (p=0.008 n=5+5)

name           old alloc/op   new alloc/op   delta
_Serialize-10    3.94MB ± 0%    0.13MB ± 0%  -96.59%  (p=0.008 n=5+5)

name           old allocs/op  new allocs/op  delta
_Serialize-10     40.2k ± 0%      1.2k ± 0%  -97.04%  (p=0.008 n=5+5)
```