@asubiotto 2023-03-07
> > Nice, we could probably also check each part's compaction level and avoid merging those as well since level 1 parts are non-overlapping with each other. This makes me think, should we maybe force a compaction of the table block to have only level 1 parts and then serialize those or would that be wasting CPU?
> 
> They're non-overlapping but are they ordered? I think I'll open up an issue around this improvement.

No, but you can sort them with the `PartSorter`.